Save position in all windows with buffer
Originally, this logic only saved the position in the current window
for the buffer. If the same buffer is open in multiple windows, all
the other windows would jump to the beginning of the buffer. This
can be rather distracting when working on a complex module.

With this change, the buffer position is saved and restored for all
windows with the affected buffer.

Fix #49